refactor: split open file tab button
